SHF-16: integrate user type to admin
